+++ AGENTS.md
@@ -73,7 +73,11 @@ qq-maid-common / reqwest / serde / tokio
 - 修改已有逻辑时同步检查附近注释是否仍准确；只有注释明显错误、重复或失去意义时才删除。
 - 不要把具体人设、群聊内容、真实用户信息或业务材料写死进代码。
 - 修改文档时避免复制 README 大段细节，优先链接到已有权威文档。
-- 代码尽量不要超过1000行，函数不要超过100行，单个文件不要超过3个模块；必要时拆分到 `mod.rs` + `<submodule>.rs`。
+- 单文件达到约 1000 行时，应检查是否存在可独立命名的职责边界；1000 行是触发架构审查的信号，不是强制验收阈值，不得仅为满足行数限制机械移动代码。
+- `mod.rs` 优先作为模块门面、核心类型和子模块编排入口，不应单纯为了规避行数限制承载接近 1000 行的混合实现。
+- 避免使用语义模糊的大型 `helpers.rs` / `utils.rs` 承接拆出的杂项；逻辑可以按领域职责命名时，应优先使用 `session.rs`、`bootstrap.rs`、`addressing.rs` 等明确名称。
+- 不是所有 700～900 行文件都必须继续拆分；只有职责明显混杂时才需要进一步整理。强相关逻辑可以保留在同一文件中，不要为了数字指标破坏局部可读性。
+- 函数尽量不要超过 100 行，单个文件不要超过 3 个职责模块；需要拆分时优先采用 `mod.rs` + 职责明确的子模块。
 
 ### 日志规范
