+++ qq-maid-core/src/management/auth/tests.rs
@@ -1,3 +1,5 @@
+use std::path::{Path, PathBuf};
+
 use super::*;
 use crate::storage::APP_MIGRATIONS;
